requirements.txt
selenium==4.34.2
webdriver-manager==4.0.2
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.2
